% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alignSequenceGraph}
\alias{alignSequenceGraph}
\title{Get the alignment table from a sequence to a graph.}
\usage{
alignSequenceGraph(sgrap, query)
}
\arguments{
\item{sgrap}{The sequence graph in question.}

\item{query}{The sequence to align.}
}
\value{
The alignment table.
}
\description{
Get the alignment table from a sequence to a graph.
}
