% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{approximateNetworkStats}
\alias{approximateNetworkStats}
\title{A heuristic to estimate some network properties of PCR amplified sequences}
\usage{
approximateNetworkStats(seqs, counts, leafStats, degree,
  minCoverage = 10L, ignoreHomopolymers = FALSE)
}
\arguments{
\item{seqs}{(these are unique sequences found at some loci)}

\item{counts}{(parallel array to seqs; the number of times a particular allele was observed)}

\item{leafStats}{(leaf indexes-- see description. This parallel to seqs}

\item{degree}{(same structure as leafStats, but estimates the degree of each allele)}

\item{ignoreHomopolymers}{(consecutive sequences of the same letter is treated as just 1 instance of that letter)}
}
\description{
This function uses a heuristic to 
infer some very basic network properties on a bunch of sequences
We define a leaf to be some sequence with distance < 3 to some other sequence in the set
a sequence is a leaf iff:
If this is distance 1, then exactly one sequence is distance 1 away
If this is distance 2, then only one sequence is distance 2 away
if this property holds then the leafStats vector is populated with the index
of the allele (some index i in seqs where seqs[i] is the node that this allele is uniquely
connected to.
The second statistic is the degree. In networks the degree of some node (sequence in our case)
is the number of edges connected to that node. For us, it's the number of sequences where that
sequence is the argmin of the distance function (aka, a reverse nearest neighbor). 
As there may be ties, if some sequence is, say, distance 1 from two sequences, the 
node with the higher count (and lowest order amongst equal counts) is selected.
all of these network properties are assuming that the distance is small (Hamming < 3)
greater distances are evaluated separately.
}
