% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypical.R
\name{bedBamAndBeyond}
\alias{bedBamAndBeyond}
\title{bed+bam->haplotypes}
\usage{
bedBamAndBeyond(bedFile, bamFile, ...)
}
\arguments{
\item{bedFile}{(A file in bed-format)}

\item{bamFile}{(an indexed BAM file)}

\item{...}{(additional arguments passed to Rsamtools::ScanBamParam)}
}
\description{
This takes a bunch of genomic intervals in BED format
and a BAM file, and it extracts every haplotype
from each genomic region listed...
}
