% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimateHaplotypes}
\alias{estimateHaplotypes}
\title{Alignment-assisted STRait Razor in R}
\usage{
estimateHaplotypes(
  seqs,
  cigars,
  seqStarts,
  qWidths,
  strands,
  starts,
  stops,
  regionStart,
  regionStop
)
}
\arguments{
\item{seqs}{(query sequences)}

\item{cigars}{(query cigar operations (as per the Bam file format))}

\item{seqStarts}{(the base position in 1-based indexing of the first base in the alignment)}

\item{qWidths}{(the size or width of the alignment in the reference genome)}

\item{strands}{(the size or width of the alignment in the reference genome)}

\item{starts}{(of the regions you're intersted in, the 0-based start coordinate)}

\item{stops}{(of the regions you're intersted in, the 1-based stop coordinate)}

\item{regionStart}{(currently ignored; specifies the amplicon start/stop)}

\item{regionStop}{(currently ignored; specifies the amplicon start/stop)}
}
\description{
This is an alignment-assisted version of STRait Razor. See:
<https://doi.org/10.1016/j.fsigen.2013.04.005>
and for shameless self promotion, see:
<https://doi.org/10.1016/j.fsigen.2017.05.008>
It is used to query a given locus, usually a short tandem repeat, and for it to return the variation in and around the repeat.
It is typically applied to DNA that has been PCR amplified.
STRait Razor solves this problem by just grabbing the sequence around where varaition is known to exist. 
The previous versions used "anchors" (e.g., primers) to grab the sequences associated with a known locus
This version uses genomic coordinates. You can optionally query for the entire genomic region, which gives you functionally
equivalent allele calls as the prevous versions of STRait Razor. You can optionally also mask out regions (troublesome homopolymers)
or you can use it to associate SNP calls in the flanking regions (which are well-aligned) with the STR region itself.
Or even to evaluate the co-association of alleles (e.g., physical phase)
As input the algorithm expects DNA sequences (seqs) 
that have been aligned (cigars)
to a reference genome (seqStarts)
And a set of queries are given (starts, stops)
and all reads that span all starts,stops are assessed
and the co-association of alleles in these blocks is assessed.
}
