% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastBoundedHammingGraphDist}
\alias{fastBoundedHammingGraphDist}
\title{Search a Hamming graph}
\usage{
fastBoundedHammingGraphDist(hammingGraph, toCompare, maxDist)
}
\arguments{
\item{hammingGraph}{(string (singular) from makeSequenceHammingGraph)}

\item{toCompare}{(vector of strings; must be over set {ACGT})}

\item{maxDist}{(set to <0 if no bound; otherwise the computation halts when maxDist differences found)}
}
\description{
Returns vector of graph distances 
in particular, the distance of each sequence (toCompare)
to the graph using the Hamming distance
}
