% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastBoundedHammingRange1NN}
\alias{fastBoundedHammingRange1NN}
\title{Quick UMI NN search}
\usage{
fastBoundedHammingRange1NN(umis, counts, tolerance = 0L)
}
\arguments{
\item{umis}{(vector of strings (UMI sequences))}

\item{counts}{(parallel vector of counts;the $ of times (reads) that the UMI was observed in)}

\item{tolerance}{(when 0, equal or more abundant is the criterion; l allows for "equal" to include 1-less)}
}
\description{
This takes a vector of UMIs and counts ; these must be sorted in descending order
It traverses the UMIs, and gives the index (1-based) for any UMI in which there is another
UMI that is more (or equally) abundant AND 1 nucleotide apart
Values of 0 correspond to no UMI that is more abundant and 1 nuc separated.
}
