% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastCloseDistances}
\alias{fastCloseDistances}
\title{Computes fast string distances when the strings are close.}
\usage{
fastCloseDistances(query, seqs, result, ignoreHomopolymers = FALSE)
}
\arguments{
\item{query}{(query sequence)}

\item{seqs}{(target sequences)}

\item{ignoreHomopolymers}{(consecutive sequences of the same letter is treated as just 1 instance of that letter)}
}
\description{
This function computes a bounded, naive (and fast)
Hamming distance between query and bunch of sequence targets (seqs)
It is only designed to be correct iff that distance < 3
and may use short circuit evaluation when the distance is >= 3
It optionally ignores homopolymers (e.g., AAAT == AT == distance of 0)
this version is vectorized (on seqs)
}
