% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastCloseHammingPair}
\alias{fastCloseHammingPair}
\title{Naive and fast Hamming distance (max dist of 2)}
\usage{
fastCloseHammingPair(query, target, ignoreHomopolymers = FALSE)
}
\arguments{
\item{query}{(query sequence)}

\item{target}{(target sequence)}

\item{ignoreHomopolymers}{(consecutive sequences of the same letter is treated as just 1 instance of that letter)}
}
\description{
This function computes a naive (and fast)
Hamming distance between query and target
It is only designed to be correct iff that distance < 3
and may use short circuit evaluation when the distance is >= 3
It optionally ignores homopolymers (e.g., AAAT == AT == distance of 0)
}
