% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAlignmentMissedNodes}
\alias{getAlignmentMissedNodes}
\title{Note which nodes of the graph were not visited in an alignment.}
\usage{
getAlignmentMissedNodes(sgrap, alnTable)
}
\arguments{
\item{sgrap}{The original graph.}

\item{alnTable}{The result of the alignment.}
}
\value{
The nodes of the graph which were not visited.
}
\description{
Note which nodes of the graph were not visited in an alignment.
}
