% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypical.R
\name{getMultiHaplotypePileup}
\alias{getMultiHaplotypePileup}
\title{Generates a multi-locus haplotypes}
\usage{
getMultiHaplotypePileup(
  chrom,
  b,
  e,
  bamFilename,
  bed = TRUE,
  makedf = TRUE,
  mapqFilter = 3,
  ...
)
}
\arguments{
\item{chrom}{(chromosome, really a chromosome/contig ID from the reference mapped to)}

\item{b}{(begin coordinate (vector))}

\item{e}{(end coordinate (vector))}

\item{bamFilename}{(the name of the bam file)}

\item{bed}{(are the locus in BED format (0-based b, 1-based e)? If set to FALSE 1-based indexing is assumed)}

\item{makedf}{(when FALSE, returns a list; each element corresponds to each locus; otherwise it returns a tibble)}

\item{mapqFilter}{(minimum mapping quality to consider...)}

\item{...}{(additional arguments passed to ScanBamParam)}
}
\description{
This extracts ALL haplotypes from ALL regions specified
chrom, begin and end specify the locus coordinates
and the bamFilename is the STRING sequence of the bam FILE
(not an Rsamtools object)
if madedf==TRUE then 1 data frame is returned with:
The Locus, the Haplotype, the PlusCounts and the NegCounts
else it returns a LIST of dataframes (1 for each Locus)
}
\details{
additional arguments to this function are passed to ScanBamParam
(letter you specify what filtering you would prefer)
}
