% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypical.R
\name{getReferenceSubstring}
\alias{getReferenceSubstring}
\title{gets reference fragments}
\usage{
getReferenceSubstring(s, chrom, b, e, bed = TRUE)
}
\arguments{
\item{s}{(seqinr object)}

\item{chrom}{(chromosomes/scaffolds/fasta identifiers)}

\item{b}{(begin coordinate, 0-based)}

\item{e}{(end coordinate, 1-based)}

\item{bed}{(is bed format? when FALSE 1-based indexing is assumed)}
}
\description{
Takes in an seqinr object and returns a character string for 
a given "chromosome" (including 'chrM'), and coordinates
(begin=b, end=e) in the reference genome. It returns
the sequences associated with these regions (as a vector).
}
\details{
The implementation is vectorized (c, b and e may be of length>1)
}
