% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSequenceGraphEditDistance}
\alias{getSequenceGraphEditDistance}
\title{Get the edit distance from a sequence to a graph.}
\usage{
getSequenceGraphEditDistance(alnTable)
}
\arguments{
\item{alnTable}{The result of the alignment.}
}
\value{
The edit distance between the graph and the query.
}
\description{
Get the edit distance from a sequence to a graph.
}
