% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypical.R
\name{getSingleHaplotypePileup}
\alias{getSingleHaplotypePileup}
\title{Generates a single-locus haplotypes}
\usage{
getSingleHaplotypePileup(
  chrom,
  b,
  e,
  bamFilename,
  bed = TRUE,
  mapqFilter = 3,
  ...
)
}
\arguments{
\item{chrom}{(chromosome, really a chromosome/contig ID from the reference mapped to)}

\item{b}{(begin coordinate (vector))}

\item{e}{(end coordinate (vector))}

\item{bamFilename}{(the name of the bam file)}

\item{bed}{(are the locus in BED format (0-based b, 1-based e). If set to FALSE 1-based indexing is assumed)}

\item{mapqFilter}{(minimum mapping quality to consider...)}

\item{...}{(additional arguments passed to ScanBamParam)}
}
\description{
This extracts ALL haplotypes from 1 region specified
chrom, begin and end specify the locus coordinate
and the bamFilename is the STRING sequence of the bam FILE
(not an Rsamtools object)
The the Haplotype, the PlusCounts and the NegCounts
are returned (as a data frame)
}
