% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{graphToString}
\alias{graphToString}
\title{Prepare a string representation of the graph.}
\usage{
graphToString(sgrap)
}
\arguments{
\item{sgrap}{The sequence graph in question.}
}
\value{
A string representation of the graph.
}
\description{
Prepare a string representation of the graph.
}
