% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypical.R
\name{haplotypeHelper}
\alias{haplotypeHelper}
\title{Extacts a single locus' haplotypes}
\usage{
haplotypeHelper(bam, b, e)
}
\arguments{
\item{bam}{(Rsamtools object: from scanBam, singular)}

\item{b}{(begin coordinate (singular))}

\item{e}{(end coordinate (singular!))}
}
\description{
*This is a helper function* Do not use unless you know what you're doing!
}
\details{
This extracts the haplotypes associated with a single locus
The locus is pre-extracted (bam object) and the begin (b)
and end (e) coordinate is specified.
}
