% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ksw2_gg_align}
\alias{ksw2_gg_align}
\title{Global pairwise alignment from the ksw2 library}
\usage{
ksw2_gg_align(Tseq, Qseq, opPos, ops, sc_mch = 1L, sc_mis = -2L,
  gapo = 2L, gape = 1L, extended = TRUE)
}
\arguments{
\item{Tseq}{(a string from the DNA alphabet; this is your target)}

\item{Qseq}{(like Tseq, but this is your query)}

\item{opPos}{(integer vector that is modified. It is the position of the cigar operation)}

\item{ops}{(these are the cigar operations themselves. there encoded as their integer representations, e.g., int('M')}

\item{sc_mch}{(the score for a match)}

\item{sc_mis}{(the penalty for a mismatch)}

\item{gapo}{(gap open penalty)}

\item{gape}{(gap extend penalty)}

\item{extended}{(changes M [Match or Mismatch] to =/X [Match or Mismatch, respectively] in the CIGAR)}
}
\description{
This performs global pairwise alignment between a pair of sequences.
It returns the number of cigar operations
and the operations (M, I, D, optionally =/X if extended CIGARs are used)
(ops vector)
and the operation positions (opPos vector)
}
\details{
Written by Heng Li with small tweaks by August Woerner
}
