% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ksw2_seqs2seqdiffs}
\alias{ksw2_seqs2seqdiffs}
\title{Generates a diff between two sequences.}
\usage{
ksw2_seqs2seqdiffs(Tseq, Qseq, sc_mch = 1L, sc_mis = -2L, gapo = 2L, gape = 1L)
}
\arguments{
\item{Tseq}{(the target sequence; e.g., the reference genome)}

\item{Qseq}{(the query sequence)}

\item{sc_mch}{(the score for a match)}

\item{sc_mis}{(the penalty for a mismatch)}

\item{gapo}{(gap open penalty)}

\item{gape}{(gap extend penalty)}
}
\description{
This is written by Heng Li with small tweaks by August Woerner
It performs global alignment (parameterized below)
and returns a data frame  of the sequence differences
}
