% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makeSequenceHammingGraph}
\alias{makeSequenceHammingGraph}
\title{Make a Hamming graph}
\usage{
makeSequenceHammingGraph(refRS, diffAllele, positions)
}
\arguments{
\item{refRS}{(The reference sequence)}

\item{diffAllele}{(vector of strings; must contain only single nucleotides)}

\item{positions}{(vector if integers; must contain the positions)}
}
\description{
Conceptually, it takes a reference sequence and adds IUPAC
codes to it where substitutions occur
}
\details{
diffAllele encodes the sequence differences
and
positions give the index of the position

The actual encoding uses bit-operations instead, but the take-home is the same
Queries are then allowed to match any of the letters in the IUPAC encoding
}
