% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{partitionBed}
\alias{partitionBed}
\title{Partitions a bed file.}
\usage{
partitionBed(chromsQ, qStarts, qStops, chromsT, tStarts, tStops)
}
\arguments{
\item{chromsQ}{(query chromosome)}

\item{qStarts}{(query start coordinate. 0-based indexing)}

\item{qStops}{(query stop coordinate. 1-based indexing)}

\item{chromsT}{(target chromosome)}

\item{tStarts}{(target start coordinate. 0-based indexing)}

\item{tStops}{(target stop coordinate. 1-based indexing)}
}
\description{
This function takes two genomic intervals: one for a query (Q)
and another for the target (T).
The genomic intervals are of the form: chromosome, start, stop
which specifies a genomic range. Genomic ranges are in half-open coordinates.
See: http://genome.ucsc.edu/blog/the-ucsc-genome-browser-coordinate-counting-systems/
This creates a dataframe of the genomic intersection: every element in Q
is paired with every overlapping element in T.
This can be useful for associating a set of SNPs, say, with the amplicons in which they
arose.
}
