% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypical.R
\name{readBed}
\alias{readBed}
\title{reads BED format}
\usage{
readBed(bedfile, ...)
}
\arguments{
\item{bedfile}{(name of bed file)}

\item{...}{(xtra arguments passed to readr::read_tsv)}
}
\description{
Reads genomic regions in BED format
https://genome.ucsc.edu/FAQ/FAQformat.html#format1
}
\details{
Only the first 3 columns are considered
Further, the first 3 columns are defined as: chrom, start, stop
and start and stop are forced into integers...
}
