% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypical.R
\name{readReferenceFasta}
\alias{readReferenceFasta}
\title{Reads reference fasta}
\usage{
readReferenceFasta(f, seqonly = FALSE, ...)
}
\arguments{
\item{f}{(fasta file)}

\item{seqonly}{(whether to give just strings (TRUE) or seqinr objects (FALSE))}

\item{...}{(passed to seqinr::read.fasta)}
}
\description{
This reads in a the reference sequence (fasta)
it does so by wrapping the seqinr::read.fasta function.
This has been hard coded for a DNA alphabet and
to get character strings
}
