% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seqdiffs2seq}
\alias{seqdiffs2seq}
\title{String reconstruction from a difference encoding...}
\usage{
seqdiffs2seq(Tseq, positions, types, events, initBuff = -1L)
}
\arguments{
\item{Tseq}{(the target sequence; e.g., the reference genome)}

\item{positions}{(1-based positions where the sequence differences are)}

\item{types}{(the types of events (0,1,2 for mismatch, deletions and insertions, respectively)}

\item{events}{(the nucleotides involved with the event; ignored if deletion)}

\item{initBuff}{(guess as to the final size of the query sequence. Overestimating is better than under)}
}
\description{
This takes the output from: ksw2_gg_align_df(Tseq, Qseq, ... )
and the reference (Tseq)
and uses it to re-create Qseq.
The main advantage of this approach is that particular types
of mutations can be filtered, as well as variants that fall in
particular regions.
}
